Correções de rotas, controle de validação

+++ controllers/RoomController.php
@@ -1,12 +1,18 @@
 <?php
 require_once __DIR__ . "/../models/RoomModel.php";
+require_once "ValidatorController.php";
+
 
 class RoomController{
+    public static $labels = ["nome", "numero", "qtd_casal", "qtd_solteiro", "preco", "disponivel"];
 
     public static function create($conn, $data){
 
-        if( ! isset($data['disponivel']) ){
-            return jsonResponse(['message'=>"Erro, Falta o campo: disponivel"], 400);
+        $validar = ValidatorController::validate_data($data, self::$labels);
+
+        if( !empty($validar) ){
+            $dados = implode(", ", $validar);
+            return jsonResponse(['message'=>"Erro, Falta o campo: ".$dados], 400);
         }
 
 
@@ -18,8 +24,6 @@ class RoomController{
         }
     }
 
-
-
     
     public static function getAll($conn){
         $roomList = RoomModel::getAll($conn);
@@ -27,11 +31,18 @@ class RoomController{
     }
 
     public static function getById($conn, $id){
+        if( empty($id) ){
+            return jsonResponse(['message'=>"Erro, Falta o campo: id"], 400);
+        }
         $buscarId = RoomModel::getById($conn, $id);
         return jsonResponse($buscarId);
     }
 
     public static function delete($conn, $id){
+        if( empty($id) ){
+            return jsonResponse(['message'=>"Erro, Falta o campo: id"], 400);
+        }
+
         $result = RoomModel::delete($conn, $id);
         if ($result){
             return jsonResponse(['message'=>"Quarto excluido com sucesso"]);
@@ -46,6 +57,15 @@ class RoomController{
             return jsonResponse(['message'=> 'Quarto atualizado com sucesso']);
         }else{
             return jsonResponse(['message'=> 'Deu merda'], 400);
+        }
+    }
+    
+    public static function get_available($conn, $data){
+        $result = RoomModel::get_available($conn, $data);
+        if($result){
+            return jsonResponse(['Quartos'=> $result]);
+        }else{
+            return jsonResponse(['message'=> 'asdasdasd'], 400);
         }
     }
